Create doubleINTARGSfunction11.14.23.15:22
Linear Search
With a sequential or linear search, we begin at the beginning of the array and look at each element of the array until we find the key we are looking for or until we reach the end of the array (whichever comes first). 

Look at every element
Because we are looking at every element in the array one at a time, we will need a loop.  We also need to compare the needle to each element of the array, so we will need to embed an if statement in that loop to do the comparison. 

Note the linear search algorithm:  

Use a loop to get each item
Embed an if statement to compare the item in the current loop position (i) to the search key and return true if there is a match. This will end the loop and the function.
If the loop completes, that means there is no match, so return false.
This is a very straightforward loop comparing every element in the array with the key. As soon as an equal value is found, it returns. If the loop finishes without finding a match, the search failed and -1 is returned.

Performance
For small arrays, linear search is a good solution because it's so straightforward. In an array of a million elements linear search on average will take 500,000 comparisons to find the key. 

 

Here is what the sequential search function would look like:

int sequentialSearch(double haystack[], double needle, int size)
{

   for (int i = 0; i < size; i++)
   {
       if (needle == haystack[i])
       {
           return i; //i represents the index number where you find the needle
        }

    }

   return -1; //reached the end of the array and needle not found

}

 

Write the code for your client (main) function
The example below requires that you #include the iostream class and using namespace std;

  

    //Main function 
    int main() 
    { 
        
        int bananas[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10}; 
         //Search for 8 
        
        cout << "Found 8 in location #: " << sequentialSearch(bananas, 8, 10) << endl;

       return 0;
    }