Create ARRAY. 11.14.23.17:13
#include <random>
int arrSize;
cin >> arrSize;
mt19937 gen(rd());
uniform_int_distribution <> dis(1,100);
for (int i=0; i < arrSize; i++) {
  arr[i] - dis(gen);
int max = arr[0];
int min = arr[0];
for int i=1 where i < arrSize; i++) {
  if (arr[i] > max) {
    max = arr[i];
  else if (arr[i] < min) {
    min = arr[i];
*****************
Enter the size of the array: 15
Randomly generated array: 74 15 14 9 36 41 12 45 14 11 57 16 82 33 75 
Maximum element: 82
Minimum element: 9
***************************************
what does this line mean....for (int i=0; i < arrSize; i++) {
 arr[i] - dis(gen); in the preivious code. why is dis(gen) subtracted from arr[i] and what is the significance of this coded expression
View other drafts













Sure, I can help you with that.

The line of code for (int i=0; i < arrSize; i++) { arr[i] - dis(gen); } is part of a larger C++ program that finds the maximum and minimum element in an array. This line of code iterates through the array and subtracts the value of dis(gen) from each element of the array.

Here's a breakdown of the code:

for (int i=0; i < arrSize; i++): This is a for loop that iterates over the elements of the array. The loop variable i is initialized to 0, and it increments by 1 on each iteration of the loop. The loop condition i < arrSize ensures that the loop continues to iterate until all elements of the array have been processed.
arr[i]: This is an array access expression that references the element of the array at index i. In other words, arr[i] represents the value of the ith element of the array.
- dis(gen): This is a subtraction expression that subtracts the value of dis(gen) from the value of arr[i]. The dis(gen) function is a random number generator that generates random numbers between 1 and 100.
The significance of this coded expression is that it effectively shifts the values of the array elements by a random amount. This can be useful for tasks such as data normalization or randomization.