Create MAX.MIN.ARRAY.11.14.23.17:20
Searching for the Max Value in an Array
One application for the linear search algorithm is to find the maximum value in an array.  The initial steps are basically the same as searching for some search key:

Loop from the first element in the array to the last element using a loop
Compare each element in the array to a variable in an embedded if clause
However, instead of testing for equality, we test for inequality and unlike a search for a search key, we will not exit the search early.  Instead, we will iterate through every element in the array no matter what we find.

/============================================= max
// Returns the maximum value in an array or returns -1 if the array is empty.
float max(float a[], int size) {
    if(size == 0)
          return -1;
    float maxVal = a[0];     
    for (int i=1; i<size; i++) {
        if (a[i] > maxVal) {
            maxVal = a[i];
        }
    }
    return maxVal;
}//end max
The variable maxVal is declared before the loop begins. This makes the variable accessible anywhere in the function (inside and outside of the loop). The maxVal variable will initially hold whatever value is in the array at position 0. That value might be the maximum value in the array, or it might not, but it gives us a starting point.

The loop is a standard loop for traversing any array, except that it initializes the i counter variable to 1 instead of 0. This is because we are already using position 0 as our maxVal. In the body of the loop, we embed the if statement to test for inequality. If the value in position i of the array is greater than maxVal, it could potentially be the maximum value in the array, so we want to store it in maxVal.

 if (a[i] > maxVal) {
            maxVal = a[i];
Because there is no other conditional clauses, when a[i] is less than maxVal, there is nothing to do in the body of the loop, so the loop continues with the next value in the array with i++. After the loop completes, the largest value in the array will be stored in the maxVal variable, so after the closing curly brace of the loop, the function returns maxVal.

The complete steps for finding the maximum value in an array include:

Loop from the first element in the array to the last element using a loop
Compare each element in the array to a variable in an embedded if clause
If the element in the array is greater than the variable value, set the variable equal to that element in the array so the variable will always hold the larger value.
After the loop ends, return the variable.
Programming Challenge
Follow the same steps with one slight modification to find the minimum value in an array.

Max Index

Another approach to the maximum is to return the index of the maximum value instead of the value. This is an advantage when the values are large or contain dynamically allocated values that makes assignment a non-trivial operation. Another reason to return the index is so that the value at that location in the array can be changed. To do this, we make one slight change to the above example on line 9.  Instead of setting maxVal equal to a[i], we set maxVal equal to i. Remember, i represents the current index number in the loop, so setting maxVal equal to i saves the index number that contains the maximum value in the maxVal variable.  
Max Index
  1 
  2 
  3 
  4 
  5 
  6 
  7 
  8 
  9 
 10 
 11 
 12 
//============================================= maxIndex
// Returns the index of the maximum value in an array or returns -1 if the array is empty.
int maxIndex(float a[], int size) {
    if (size > 0)
         return -1;
    int maxIndex = 0;
    for (int i=1; i<size; i++) {
        if (a[i] > a[maxIndex]) {
            maxIndex = i;
        }
    }
    return maxIndex;
}//end maxIndex
Programming Challenge
Follow the same steps with one slight modification to find the index number that contains the minimum value in an array.