Create swap.11.15.23.18:10
what do i need the temp variable for in the swap algorithm...could i just not swap the two indices without temporarily storing them in the temp variable in the program...#include <string>
#include <iostream>

using namespace std;

int main() {

 string myData = "ABCDEFGH";
 
 char temp = myData[2];
 myData[2] = myData[5];
 myData[5] = temp;
 cout << myData << endl; 
}
View other drafts






















The temp variable is crucial in the swap algorithm because it allows you to temporarily store the value of one variable while you assign the value of the other variable to it. Without the temp variable, you would overwrite the original value of one of the variables before you have a chance to store it elsewhere.

In the provided code, you're swapping the values of myData[2] and myData[5]. Without the temp variable, you would end up overwriting the value of myData[2] before assigning it to myData[5]. As a result, the original value of myData[2] would be lost.

The temp variable solves this problem by providing a temporary storage location for the value of myData[2] while you assign the value of myData[5] to it. This allows you to safely swap the values of the two variables without losing any data.

Here's a breakdown of the code:

Store myData[2] in temp: The first step is to store the value of myData[2] in the temp variable. This ensures that you have a backup of the original value before you overwrite it.

**Swap myData[2] and myData[5]: Next, you assign the value of myData[5] to myData[2]. This effectively overwrites the original value of myData[2].

**Assign temp to myData[5]: Finally, you assign the value of temp to myData[5]. This restores the original value of myData[2] to myData[5], completing the swap operation.

Without the temp variable, the swap would not be possible, and you would lose the original value of myData[2]. Therefore, the temp variable plays a critical role in ensuring that the swap operation is performed correctly.