Create Swap...11.16.23.15:59
Use a loop to cout the contents of the array: once before calling the functions and then again, after calling the functions. to test the function.  ...for....void swapShellsFirstInArray(int basket[5])
{
int temp = basket[0];
basket[0] = basket[1];
basket[1] = temp;
}

int collection[5] = {1, 2, 3, 4, 5};
swapShellsFirstInArray(collection);
C++
#include <iostream>

using namespace std;

void swapShellsFirstInArray(int basket[5])
{
  int temp = basket[0];
  basket[0] = basket[1];
  basket[1] = temp;
}

int main()
{
  int collection[5] = {1, 2, 3, 4, 5};
  cout << "Before swapping: " << endl;
  for (int i = 0; i < 5; i++)
  {
    cout << collection[i] << " ";
  }
  cout << endl;

  swapShellsFirstInArray(collection);

  cout << "After swapping: " << endl;
  for (int i = 0; i < 5; i++)
  {
    cout << collection[i] << " ";
  }
  cout << endl;

  return 0;
}
Use code with caution. Learn more
This code will output the following:

Before swapping:
1 2 3 4 5

After swapping:
2 1 3 4 5
As you can see, the function swapShellsFirstInArray correctly swaps the first two elements of the array.

profile picture
if i++ is called an increment operator, what is i < 5 called, and what is int i = 0 called in...for (int i = 0; i < 5; i++)
View other drafts


















In the for loop for (int i = 0; i < 5; i++), the following statements have specific roles:

int i = 0: This statement is called the initialization statement. It declares and initializes the loop control variable i to the initial value of 0.

i < 5: This statement is called the condition expression. It determines whether the loop should continue iterating. The loop continues as long as the value of i is less than 5.

i++: This statement is called the update statement. It modifies the loop control variable i after each iteration of the loop. In this case, it increments the value of i by 1, effectively counting up from 0 to 4.

Therefore, i++ is called the increment operator, i < 5 is called the condition expression, and int i = 0 is called the initialization statement in the for loop for (int i = 0; i < 5; i++).