Create QUIZLET.11.17.23.15:19
Consider this line of code: sort(str.begin(), str.end());

Q1: 

What is the purpose of the sort function in this line of code?

To reverse the order of the characters in the string str
To remove duplicate characters from the string str
To sort the characters in the string str in alphabetical order
To find the length of the string str

Answer:

The sort function is used to sort the characters in the string str in alphabetical order.




******************************************************************************

 Q2: WHERE WOULD I FIND THE sort() function in the c++ standard library?

The sort function is defined in the <algorithm> header file of the C++ Standard Library.

https://en.cppreference.com/w/cpp/header/algorithm

*************************************************************************

Q3. CONSIDER A SWAP ALGORITHM

What is the maximum number of comparisons that binary search will make to find a value in a sorted array?

O(1)
O(log n)
O(n)
O(n^2)
The maximum number of comparisons that binary search will make to find a value in a sorted array is O(log n), where n is the length of the array.
**************************************************
Q4. 

How does the std::sort function compare strings?

By comparing the lengths of the strings
By comparing the ASCII codes of the characters in the strings
By comparing the lexicographical order of the strings
By comparing the memory addresses of the strings

Answer:

The std::sort function compares strings by comparing the lexicographical order of the characters in the strings. This means that the strings are compared one character at a time, and the order of the strings is determined by the alphabetic order of the characters.




*****************************************************************************************************************. 

Q5. LINEAR SEARCH:

When would you typically use a linear search algorithm?

When the data is already sorted
When the data is very large
When the data is small and unsorted
When the data is sparse

Answer:

Linear search is most efficient when the data is small and unsorted. This is because linear search does not require the data to be sorted, and it only needs to compare each element of the array to the search key until it finds the desired value or reaches the end of the array.